Add timezone setting to admin interface
There were 2 options for implementing this. We chose option 1.

(1) One timezone per Django install. Django runs in this timezone
    using timezone.activate() and every site on a multi-site install
    must be the same

(2) One timezone per site. For this to work Django should run in the UTC
    timezone, but we would have to adjust every datetime display to
    ensure it takes account of a new SiteSetting.

As this isn't a SiteSetting, I've created a CmsSetting to store it
separately. Unfortunately Wagtail doesn't have the concept of settings
which apply to the install, so I've overridden the save() method to
apply the same timezone to all the other sites.

The AppConfig reads from the database at startup and sets the correct
timezone. If the timezone is null in the database, we default to the
value in the settings. If the database is not available (like in a new
app) we default to the value in the settings.

The migration runs some Python code:

- Seed initial list of timezones based on pytz common timezones
- Set the initial timezone value for each site to the value in
  Django settings