Use Wagtail hook to delete page translation
Previously we used Django receivers to delete page translations
whenever a page was deleted, but this is complex and relies on
a global variable `pages_to_delete`.

It's much better to use a Wagtail hook for this. Unfortuantely there's
still a bit of complexity here. The before_delete_page hook runs
for all requests (GET and POST, on the delete confirmation page).
The after_delete_page runs after the page has been deleted, so we
can't fetch the translations.

I think the best solution is to run this in the before hook but to
duplicate the logic to test if the request is a POST. It's not perfect
but it's nicer than the Django signals.