Remove get_image_impression_hash function
This was deliberately kept in d4be6c570a44a652e0d50fecc35d0350e4cf78a5
but it's not used anywhere.

The use of StringIO breaks compatibility with Python 3 so if we want
to use this code again in the future we'd need to update it. Let's
remove it for now and see if we need it again.