Add template lint script
Projects that depend on molo can use this as part of their build process
to ensure that their templates are well-formed.

It's not completely bulletproof but it catches some things:

(1) The templates all are parsed by Template() without error
(2) There are no duplicate tag loads
(3) All of the loads are used

I was surprised that after a bit of Googling I couldn't find a package
to do this.

It's not a great solution but if it's buggy we can easily remove it.